Replaced all amounts with BigDecimal instead of the previously double type

+++ src/main/java/io/qbeat/models/TaxConfigProperty.java
@@ -1,11 +1,13 @@
 package io.qbeat.models;
 
+import java.math.BigDecimal;
+
 public class TaxConfigProperty {
     private final int rangeStartPrice;
     private final int rangeEndPrice;
-    private final double rate;
+    private final BigDecimal rate;
 
-    public TaxConfigProperty(int rangeStartPrice, int rangeEndPrice, double rate) {
+    public TaxConfigProperty(int rangeStartPrice, int rangeEndPrice, BigDecimal rate) {
         this.rangeStartPrice = rangeStartPrice;
         this.rangeEndPrice = rangeEndPrice;
         this.rate = rate;
@@ -19,7 +21,7 @@ public class TaxConfigProperty {
         return rangeEndPrice;
     }
 
-    public double getRate() {
+    public BigDecimal getRate() {
         return rate;
     }
